Relocates some of the logic for sourcing on-machine network configuration from api/common to core/network.
This is preamble to having different implementations of
(Network)ConfigSource.

More of this logic should ultimately be move out of api/common, where it
definitely does not belong.
